dataset-service/src/main/java/org/eea/dataset/io/notification/events/UpdateDatacollectionFailedEvent.java
package org.eea.dataset.io.notification.events;

import java.util.HashMap;
import java.util.Map;
import org.eea.exception.EEAException;
import org.eea.interfaces.controller.dataflow.DataFlowController.DataFlowControllerZuul;
import org.eea.kafka.domain.EventType;
import org.eea.kafka.domain.NotificationVO;
import org.eea.notification.event.NotificableEventHandler;
import org.springframework.beans.factory.annotation.Autowired;

public class UpdateDatacollectionFailedEvent implements NotificableEventHandler {

  @Autowired
  private DataFlowControllerZuul dataflowControllerZuul;

  @Override
  public EventType getEventType() {
    return EventType.UPDATE_DATACOLLECTION_FAILED_EVENT;
  }

  @Override
  public Map<String, Object> getMap(NotificationVO notificationVO) throws EEAException {
    Long dataflowId = notificationVO.getDataflowId();
    String dataflowName =
        notificationVO.getDataflowName() != null ? notificationVO.getDataflowName()
            : dataflowControllerZuul.findById(dataflowId).getName();

    Map<String, Object> notification = new HashMap<>();
    notification.put("user", notificationVO.getUser());
    notification.put("dataflowId", dataflowId);
    notification.put("dataflowName", dataflowName);
    notification.put("error", notificationVO.getError());
    return notification;
  }

}
